refactor: 순수 유틸  함수들 분리
